removed getSettings method from applicationContainer in favour of MigrationHelpers
